+++ .github/workflows/Secp_Disabled_Tests.yml
@@ -25,4 +25,4 @@ jobs:
             ~/.bitcoin-s/binaries
           key: ${{ runner.os }}-cache
       - name: run tests
-        run: sbt ++2.13.5 cryptoTest/test coreTest/test
+        run: sbt ++2.13.5 cryptoTestJVM/test coreTestJVM/test
