+++ core-test/.jvm/src/test/resources/00000000000000000008513c860373da0484f065983aeb063ebf81c172e81d48.txt
