+++ core-test/.jvm/src/test/resources/base58_encode_decode.json
