+++ core-test/.jvm/src/test/resources/base58_keys_invalid.json
