+++ core-test/.jvm/src/test/resources/bigsize_decoding.json
