+++ core-test/.jvm/src/test/resources/bigsize_encoding.json
