+++ core-test/.jvm/src/test/resources/bitcoin-util-test.json
