+++ core-test/.jvm/src/test/resources/script_tests.json
