+++ core-test/.jvm/src/test/resources/sighash.json
