+++ core-test/.jvm/src/test/resources/trezor.txt
