+++ core-test/.jvm/src/test/resources/trezor_bip39_vectors.json
