+++ core-test/.jvm/src/test/resources/tx_invalid.json
