+++ core-test/.jvm/src/test/resources/tx_valid.json
