+++ core/.jvm/src/main/resources/bip39-wordlists/english.txt
