+++ crypto-test/src/test/scala/org/bitcoins/crypto/ECDigitalSignatureTest.scala
@@ -94,14 +94,4 @@ class ECDigitalSignatureTest extends BitcoinSUnitTest {
     }
   }
 
-  it must "be able to generate valid signatures with bouncy castle" in {
-    forAll(CryptoGenerators.privateKey, CryptoGenerators.sha256Digest) {
-      case (privKey: ECPrivateKey, hash: Sha256Digest) =>
-        val sig = BouncycastleCryptoRuntime.sign(privKey, hash.bytes)
-        val pubKey = privKey.publicKey
-
-        assert(pubKey.verify(hash, sig))
-    }
-  }
-
 }
